# backend/docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Risk Check API","version":"0.1.0"},"paths":{"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/works/groups":{"get":{"tags":["works"],"summary":"List Groups","operationId":"list_groups_works_groups_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WorkGroupResponse"},"type":"array","title":"Response List Groups Works Groups Get"}}}}}},"post":{"tags":["works"],"summary":"Create Group","operationId":"create_group_works_groups_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkGroupCreateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkGroupResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/works":{"post":{"tags":["works"],"summary":"Create Work","operationId":"create_work_works_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkCreateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/works/{work_id}/items":{"post":{"tags":["works"],"summary":"Add Work Item","operationId":"add_work_item_works__work_id__items_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkItemCreateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkItemResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/items/{work_item_id}/risks/generate":{"post":{"tags":["works"],"summary":"Generate Risk Assessment","operationId":"generate_risk_assessment_works_items__work_item_id__risks_generate_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"work_item_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RiskAssessmentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/daily":{"get":{"tags":["works"],"summary":"Get Daily Overview","operationId":"get_daily_overview_works_daily_get","parameters":[{"name":"work_date","in":"query","required":true,"schema":{"type":"string","format":"date","title":"Work Date"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkOverviewResponse"},"title":"Response Get Daily Overview Works Daily Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/works/{work_id}/comments":{"post":{"tags":["comments"],"summary":"Add Comment","operationId":"add_comment_works__work_id__comments_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkCommentCreateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkCommentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["comments"],"summary":"List Comments","operationId":"list_comments_works__work_id__comments_get","parameters":[{"name":"work_id","in":"path","required":true,"schema":{"type":"integer","title":"Work Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkCommentResponse"},"title":"Response List Comments Works  Work Id  Comments Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RiskAssessmentResponse":{"properties":{"id":{"type":"integer","title":"Id"},"work_item_id":{"type":"integer","title":"Work Item Id"},"content":{"type":"string","title":"Content"},"generated_at":{"type":"string","format":"date-time","title":"Generated At"}},"type":"object","required":["id","work_item_id","content","generated_at"],"title":"RiskAssessmentResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"role":{"type":"string","title":"Role"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","name","role","is_active"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WorkCommentCreateRequest":{"properties":{"content":{"type":"string","maxLength":500,"minLength":1,"title":"Content"}},"type":"object","required":["content"],"title":"WorkCommentCreateRequest"},"WorkCommentResponse":{"properties":{"id":{"type":"integer","title":"Id"},"work_id":{"type":"integer","title":"Work Id"},"user_id":{"type":"integer","title":"User Id"},"content":{"type":"string","title":"Content"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","work_id","user_id","content","created_at"],"title":"WorkCommentResponse"},"WorkCreateRequest":{"properties":{"title":{"type":"string","maxLength":100,"minLength":1,"title":"Title"},"description":{"type":"string","maxLength":500,"minLength":1,"title":"Description"},"group_id":{"type":"integer","title":"Group Id"},"work_date":{"type":"string","format":"date","title":"Work Date"},"status":{"type":"string","title":"Status","default":"draft"}},"type":"object","required":["title","description","group_id","work_date"],"title":"WorkCreateRequest"},"WorkGroupCreateRequest":{"properties":{"name":{"type":"string","maxLength":50,"minLength":1,"title":"Name"}},"type":"object","required":["name"],"title":"WorkGroupCreateRequest"},"WorkGroupResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"WorkGroupResponse"},"WorkItemCreateRequest":{"properties":{"name":{"type":"string","maxLength":100,"minLength":1,"title":"Name"},"description":{"type":"string","maxLength":300,"minLength":1,"title":"Description"}},"type":"object","required":["name","description"],"title":"WorkItemCreateRequest"},"WorkItemResponse":{"properties":{"id":{"type":"integer","title":"Id"},"work_id":{"type":"integer","title":"Work Id"},"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"}},"type":"object","required":["id","work_id","name","description"],"title":"WorkItemResponse"},"WorkItemWithRisksResponse":{"properties":{"item":{"$ref":"#/components/schemas/WorkItemResponse"},"risks":{"items":{"$ref":"#/components/schemas/RiskAssessmentResponse"},"type":"array","title":"Risks"}},"type":"object","required":["item","risks"],"title":"WorkItemWithRisksResponse"},"WorkOverviewResponse":{"properties":{"work":{"$ref":"#/components/schemas/WorkResponse"},"items":{"items":{"$ref":"#/components/schemas/WorkItemWithRisksResponse"},"type":"array","title":"Items"}},"type":"object","required":["work","items"],"title":"WorkOverviewResponse"},"WorkResponse":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description"},"group_id":{"type":"integer","title":"Group Id"},"work_date":{"type":"string","format":"date","title":"Work Date"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","title","description","group_id","work_date","status"],"title":"WorkResponse"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}